feat : refactor match result screen to same matches package
