+++ app/src/main/java/com/karrar/movieapp/ui/match/questions/MatchQuestionInteractionListener.kt
@@ -7,4 +7,6 @@ interface MatchQuestionInteractionListener : BaseInteractionListener {
     fun onNextClicked()
 
     fun getCurrentQuestionType(): MatchQuestionType
+
+    fun onStartMatchingClicked()
 }
