--- app/src/main/java/com/karrar/movieapp/ui/match_result/MatchResultUiEvent.kt
@@ -1,9 +1,0 @@
-package com.karrar.movieapp.ui.match_result
-
-sealed interface MatchResultUiEvent {
-
-    object NavigateBack: MatchResultUiEvent
-    object ViewMovieDetails : MatchResultUiEvent
-    data class PlayYoutubeTrailer(val movieId : Int) : MatchResultUiEvent
-    data class SaveMovie(val movieId : Int) : MatchResultUiEvent
-}
